Refactor post retrieval used in the AJAX callback
This uses `get_post` instead of `wp_query`, but more importantly,
uses the `lang` query argument (provided by Polylang) instead of a
`tax_query`. For some reason, using `tax_query` to query for the
language resulted in the query failing if both the English and
Spanish page were open in the same browser session.